// src/app/models/acoes.enum.ts
export enum Acoes {
    Inicial = 'inicial',
    ProcessoSeletivo = 'processosseletivos',
    Calendario = 'calendario',
    Unidade = 'unidades',
    Curso = 'cursos',
    Turno = 'turnos',
    EstadoPolo = 'estadospolo',
    CidadePolo = 'cidadespolo',
    Polo = 'polos',
    FormaIngresso = 'formaIngresso',
    Convenio = "convenios"
  }