// src/app/models/filtro.model.ts
export class Filtro {
    codColigada: number = 1;
    codTipoCurso: number;
    categoriaPs: string;
    calendario: string;
    idCategoriaps: number;
    idPerlet: number;
    codFilial: number;
    codCurso: string;
    codTurno: number;
    idAreaInteresse: number;
    codCampus: string;
    idPs: number;
    idFormaInscricao: number;
    codBolsa: string;
    codTipoHabilitacao: string;
    formaIngresso: string;
    codPerLet: string;
    ignoraFimInscricao: boolean;
}